+++ src/lib/navbar.svelte
@@ -1,7 +1,6 @@
 <script>
 	export let data;
 	export let products;
-	import ThemeSelect from '$lib/theme-select.svelte';
 	import HorizontalNav from '../components/horizontalNav.svelte';
 
 	import { page } from '$app/stores';
@@ -18,12 +17,36 @@
 
 	let loaded = false;
 
+	const themes = ['dark', 'light'];
+
+	$: current_theme = '';
+
 	onMount(async () => {
 		title = document.title;
 		title = title;
 		notifications = await products;
+		if (typeof window !== 'undefined') {
+			const theme = window.localStorage.getItem('theme');
+			if (theme && themes.includes(theme)) {
+				document.documentElement.setAttribute('data-theme', theme);
+				current_theme = theme;
+			}
+		}
 		loaded = true;
 	});
+
+	function changeTheme() {
+		theme = '';
+		let bool = document.querySelector("#theme").checked
+		bool ? theme = 'light' : theme = 'dark';
+		if (themes.includes(theme)) {
+			const one_year = 60 * 60 * 24 * 365;
+			window.localStorage.setItem('theme', theme);
+			document.cookie = `theme=${theme}; max-age=${one_year}; path=/; SameSite=Lax`;
+			document.documentElement.setAttribute('data-theme', theme);
+			current_theme = theme;
+		}
+	}
 
 	$: path = $page.url.pathname;
 
@@ -122,34 +145,14 @@
 					</ul>
 				</div>
 			</ul>
-			<ul class="menu menu-horizontal items-center p-0">
-				<div class="dropdown dropdown-end">
-					{#if avatarUrl}
-						<label tabindex="0" class="btn btn-ghost btn-circle avatar">
-							<div class="w-10 rounded-full">
-								<img src={avatarUrl} alt="user avatar" />
-							</div>
-						</label>
-					{:else}
-						<label class="btn btn-ghost btn-circle avatar placeholder">
-							<div class="bg-secondary text-neutral-content rounded-full w-24 mx-auto lg:mx-0">
-								<span class="text-3xl">{firstName.charAt(0) + lastName.charAt(0)}</span>
-							</div>
-						</label>
-					{/if}
-					<ul
-						class="menu menu-sm dropdown-content mt-3 z-[1] p-2 shadow bg-base-100 rounded-box w-52"
-					>
-						<li>
-							<a href="/app/account" class="justify-between"> Perfil </a>
-						</li>
-						<li><a href="/auth/logout">Cerrar sesion</a></li>
-					</ul>
-				</div>
-			</ul>
+			<label class="swap swap-rotate" on:click={changeTheme}>
+				<input type="checkbox" id="theme" checked={current_theme === "light" }/>
+				<svg class="swap-on fill-current w-6 h-6" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M5.64,17l-.71.71a1,1,0,0,0,0,1.41,1,1,0,0,0,1.41,0l.71-.71A1,1,0,0,0,5.64,17ZM5,12a1,1,0,0,0-1-1H3a1,1,0,0,0,0,2H4A1,1,0,0,0,5,12Zm7-7a1,1,0,0,0,1-1V3a1,1,0,0,0-2,0V4A1,1,0,0,0,12,5ZM5.64,7.05a1,1,0,0,0,.7.29,1,1,0,0,0,.71-.29,1,1,0,0,0,0-1.41l-.71-.71A1,1,0,0,0,4.93,6.34Zm12,.29a1,1,0,0,0,.7-.29l.71-.71a1,1,0,1,0-1.41-1.41L17,5.64a1,1,0,0,0,0,1.41A1,1,0,0,0,17.66,7.34ZM21,11H20a1,1,0,0,0,0,2h1a1,1,0,0,0,0-2Zm-9,8a1,1,0,0,0-1,1v1a1,1,0,0,0,2,0V20A1,1,0,0,0,12,19ZM18.36,17A1,1,0,0,0,17,18.36l.71.71a1,1,0,0,0,1.41,0,1,1,0,0,0,0-1.41ZM12,6.5A5.5,5.5,0,1,0,17.5,12,5.51,5.51,0,0,0,12,6.5Zm0,9A3.5,3.5,0,1,1,15.5,12,3.5,3.5,0,0,1,12,15.5Z"/></svg>
+				<svg class="swap-off fill-current w-6 h-6" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M21.64,13a1,1,0,0,0-1.05-.14,8.05,8.05,0,0,1-3.37.73A8.15,8.15,0,0,1,9.08,5.49a8.59,8.59,0,0,1,.25-2A1,1,0,0,0,8,2.36,10.14,10.14,0,1,0,22,14.05,1,1,0,0,0,21.64,13Zm-9.5,6.69A8.14,8.14,0,0,1,7.08,5.22v.27A10.15,10.15,0,0,0,17.22,15.63a9.79,9.79,0,0,0,2.1-.22A8.11,8.11,0,0,1,12.14,19.73Z"/></svg>
+			</label>
 		</div>
 	</div>
-	<ul class="menu bg-base-100 w-56 h-full box fixed z-10 hidden lg:flex flex-col gap-2">
+	<ul class="menu bg-base-100 w-56 h-full box fixed z-10 hidden lg:flex flex-col gap-2 flex-nowrap overflow-scroll">
 		<HorizontalNav />
 	</ul>
 </div>
